--- package.json
@@ -1,25 +1,0 @@
-{
-  "name": "backenddev-lab5-websocket-client",
-  "version": "1.0.0",
-  "description": "",
-  "main": "app.js",
-  "scripts": {
-    "start": "node app.js",
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/DevInCube/backenddev-lab5-websocket-client.git"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "bugs": {
-    "url": "https://github.com/DevInCube/backenddev-lab5-websocket-client/issues"
-  },
-  "homepage": "https://github.com/DevInCube/backenddev-lab5-websocket-client#readme",
-  "dependencies": {
-    "express": "^4.17.1",
-    "ws": "^7.4.4"
-  }
-}
